+++ examples/skybox.c
@@ -65,7 +65,7 @@ void Draw(void)
     R3D_Begin(camera);
         for (int x = 0; x < 7; x++) {
             for (int y = 0; y < 7; y++) {
-                R3D_DrawMesh(&sphere, &materials[y * 7 + x], MatrixTranslate(x - 3, y - 3, 0.0f));
+                R3D_DrawMesh(&sphere, &materials[y * 7 + x], MatrixTranslate((float)x - 3, (float)y - 3, 0.0f));
             }
         }
     R3D_End();
