+++ examples/sprite_instanced.c
@@ -43,7 +43,7 @@ const char* Init(void)
     for (int i = 0; i < sizeof(transforms) / sizeof(*transforms); i++) {
         float scaleFactor = GetRandomValue(25, 50) / 10.0f;
         Matrix scale = MatrixScale(scaleFactor, scaleFactor, 1.0f);
-        Matrix translate = MatrixTranslate(GetRandomValue(-100, 100), scaleFactor, GetRandomValue(-100, 100));
+        Matrix translate = MatrixTranslate((float)GetRandomValue(-100, 100), scaleFactor, (float)GetRandomValue(-100, 100));
         transforms[i] = MatrixMultiply(scale, translate);
     }
